+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Fixes
 
-* Fix the Docker container paths, to make some currently broken cases work. [Stavros Korokithakis]
+* Fix the Harbormaster Docker container. [Stavros Korokithakis]
 
 * Fix issue with the Harbormaster Docker image not being able to find the data dir. [Stavros Korokithakis]