+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### Fixes
 
+* Fix tests. [Stavros Korokithakis]
+
 * Fix the Harbormaster Docker container. [Stavros Korokithakis]
 
 * Fix issue with the Harbormaster Docker image not being able to find the data dir. [Stavros Korokithakis]