+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## Unreleased
+## v0.3.2 (2023-07-23)
 
 ### Fixes